Add files via upload
Seeking help with code review or error correction. there is a problem in the website which supposed to be like a web store with toys. firstly, when I only did the html and CSS with a bit of scripts but not with complicated function it kind of worked. but when I tried to add the function all the products disappeared, and the checkout didn't work. whole the project and the functions are kind of ready. maybe the problem is with the links to the scripts. helpppp